[REF-6187] Move component's shared code to and 'extra' folder underl mlcomp
Usually, this kind of code is related to specific environment such as
'aws' or 'sagemaker', and so forth. The benefit is that the code is not
duplicated by multiple components, which were built for the given
engine/environment.
