+++ mlcomp/parallelm/common/os_util.py
@@ -1,4 +1,18 @@
+import os
 import subprocess
+import tempfile
+
+
+def tmp_filepath():
+    tmp_file = tempfile.NamedTemporaryFile()
+    filepath = tmp_file.name
+    tmp_file.close()
+    return filepath
+
+
+def remove_file_safely(filepath):
+    if os.path.isfile(filepath):
+        os.unlink(filepath)
 
 
 def service_installed(service_name):
