+++ mlcomp/parallelm/extra/sagemaker/monitor/job_monitor_transformer.py
@@ -6,9 +6,9 @@ import logging
 import pprint
 import time
 
-from monitor.job_monitor_base import JobMonitorBase
-from monitor.report import Report
-from monitor.sm_api_constants import SMApiConstants
+from parallelm.extra.sagemaker.monitor.job_monitor_base import JobMonitorBase
+from parallelm.extra.sagemaker.monitor.report import Report
+from parallelm.extra.sagemaker.monitor.sm_api_constants import SMApiConstants
 
 
 class JobMonitorTransformer(JobMonitorBase):
